requirements.txt
torch==1.8.2
torchaudio==0.8.2
torchmetrics==0.5.1
torchtext==0.9.0
torchvision==0.9.2

numpy==1.19.2
Pillow==8.3.1
Flask==1.1.2
matplotlib==3.4.1